agrovision-base-station/requirements.txt
altgraph==0.17
astroid==2.4.1
azure-core==1.6.0
azure-cosmos==4.0.0
azure-iot-device==2.1.2
azure-storage-blob==12.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
isodate==0.6.0
isort==4.3.21
janus==0.4.0
macholib==1.14
mccabe==0.6.1
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
msrest==0.6.14
numpy==1.18.4
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python-headless==4.2.0.34
paho-mqtt==1.5.0
pycparser==2.20
PyInstaller==3.6
pylint==2.5.2
PySocks==1.7.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
six==1.15.0
toml==0.10.1
tornado==4.5.3
transitions==0.8.1
typed-ast==1.4.1
uamqp==1.2.8
urllib3==1.25.9
wrapt==1.12.1
